requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-decouple==3.8
Pillow==10.4.0
reportlab==4.2.2
gunicorn==23.0.0
whitenoise==6.7.0